refactor: extract reviews feed logic into reusable ReviewFeed component
