+++ frontend/src/components/ReviewFeed/index.module.scss
@@ -1,0 +1,80 @@
+@use '../../styles/mixins' as *;
+@use '../../styles/vars' as *;
+
+.controls {
+  display: flex;
+  justify-content: space-between;
+  align-items: stretch;
+  margin: 10px 0 0;
+  gap: 20px;
+  height: 40px;
+
+  .searchForm {
+    position: relative;
+    width: 70%;
+
+    .searchInput {
+      width: 100%;
+      height: 100%;
+      padding: 8px 42px 8px 15px;
+      border: 1px solid $black;
+      box-shadow: 2px 2px 1px $black;
+      border-radius: 5px;
+      font-size: 1rem;
+      outline: none;
+      background: $my-white;
+    }
+
+    .searchButton {
+      position: absolute;
+      top: 0;
+      right: 0;
+      width: 42px;
+      height: 100%;
+      border: 1px solid $black;
+      border-radius: 0 5px 5px 0;
+      cursor: pointer;
+
+      background-color: $orange;
+      background-image: url('/search.svg');
+      background-repeat: no-repeat;
+      background-position: center;
+      background-size: 20px 20px;
+
+      &:hover {
+        background-color: lighten($orange, 5%);
+      }
+
+      &:active {
+        background-color: lighten($orange, 10%);
+      }
+    }
+  }
+
+  .sortSelect {
+    padding: 8px 15px;
+    border-radius: 5px;
+    background-color: $my-white;
+    color: $black;
+    font-size: 1rem;
+    cursor: pointer;
+    box-shadow: 2px 2px 1px $black;
+
+    &:focus {
+      outline: none;
+    }
+  }
+}
+
+.pagination {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 2rem;
+  color: $my-white;
+
+  .pageInfo {
+    min-width: 120px;
+    text-align: center;
+  }
+}
